requirements.txt
google-genai>=1.0.0
streamlit>=1.40.0
pillow>=10.0.0
python-dotenv>=1.0.0
psycopg2-binary>=2.9.0
supabase>=2.0.0
